Allow for auth headers to be constructed on a per request basis.
Prior to this commit, auth was handled at a service creation level.
This led to difficulties in a model where users are authed instead
of a service.
