+++ lib/svelte.rb
@@ -7,6 +7,7 @@ require 'svelte/configuration'
 require 'svelte/errors'
 require 'svelte/model_factory'
 require 'svelte/rest_client'
+require 'svelte/headers_builder'
 require 'svelte/service'
 require 'svelte/swagger_builder'
 require 'svelte/path'
